requirement.txt
streamlit=1.31.0
tensorflow
pandas=2.2.0
numpy=1.26.3
pillow=10.2.0